Move permission updates into SDK
